requirements.txt
pytest==8.2.1
pytest-cov==5.0.0
pylint==3.2.2
black==24.4.2
fire==0.6.0
chromadb==0.4.23
fastapi==0.110.3
Flask==3.0.3
huggingface-hub==0.22.2
langchain==0.1.16
langchain-community==0.0.34
langsmith==0.1.52
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.33
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.3
llama-index-llms-openai==0.1.16
llama-index-llms-vllm==0.1.7
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.5
llama-parse==0.4.2
llamaindex-py-client==0.1.19
mysql-connector-python==8.4.0
mysqlclient==2.1.1
ollama==0.2.0
SQLAlchemy==2.0.29
uvicorn==0.29.0
pydantic==2.7.1
pysqlite3-binary